+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.0 (2024--)
 
 - **BREAKING CHANGES**: The `md5`, `sha1`, and `sha256` methods have been renamed to `md5Hash`, `sha1Hash`, and `sha256Hash`.
+- `objMergeNewKey`: Add `objMergeNewKey` method
 
 ## 1.4.2 (2024-06-25)